Updated loft_php_lib to 0.5.7 Refactored docs to multiple pages.

+++ docs/core/markdown.php
@@ -8,12 +8,16 @@
  */
 require_once dirname(__FILE__) . '/vendor/autoload.php';
 
-use \AKlump\LoftDocs\MarkdownExtra;
+use AKlump\LoftDocs\MarkdownExtra;
+use Webuni\FrontMatter\FrontMatter;
 
-$in_file  = $argv[1];
+$in_file = $argv[1];
 $out_file = $argv[2];
 
-if ($contents = file_get_contents($in_file)) {
-  $my_html = MarkdownExtra::defaultTransform($contents);
-  file_put_contents($out_file, $my_html);
+if (is_file($in_file) && ($contents = file_get_contents($in_file))) {
+    $fm = new FrontMatter();
+    $document = $fm->parse($contents);
+    $contents = $document->getContent();
+    $my_html = MarkdownExtra::defaultTransform($contents);
+    file_put_contents($out_file, $my_html);
 }
